save other softlist details in each softlist...
...this is the first step in fixing mame's softlist mamename
uniqueness bug (where mamenames for a systems games are only unique
per-media device (so its quite possible to have smb2 mean both a floppy
and a cart for the same system). To fix we need to disambiguate at call
time with a device flag eg: 'famicom -flop1 smb2' and 'famicom -cart1
smb2'. (However muliti-disc loading (where smb2 could mean 'load 4
floppies in to the first 4 floppy devices in this order') means we only want
to do this where we have a confict)

As a first step, save out the names of other softlists when we compose
a softlist