start to write loader insertion logic
to start with i'm just trying to replicate the nes/snes replacement
code, but put it in the main json against the respective softlists.
After that i'll do a different function for the device (non-softlist)
loader insertions

A loader insertion is where a call needs like
"-cart basic -cass1 %MYROM%" ie: needs loader media to be inserted
as well as the game media.

The code here does som einsertion but they are wrong, because i'm
iterating through the reaplcement array first and then the systems json,
that needs inverting